# Simplify and defer checks in flatten to improve performance