--- a/throughs/flatten.js
+++ b/throughs/flatten.js
@@ -3,15 +3,24 @@
 var values = require('../sources/values')
 var once = require('../sources/once')
 
+function getAbortCb(read, abort, cb) {
+  return function(err) {
+    read(err || abort, cb)
+  }
+}
+function getChunkErrorCb(err, cb) {
+  return function() {
+    cb(err)
+  }
+}
+
 //convert a stream of arrays or streams into just a stream.
 module.exports = function flatten () {
   return function (read) {
     var _read
     return function (abort, cb) {
       if (abort) { //abort the current stream, and then stream of streams.
-        _read ? _read(abort, function(err) {
-          read(err || abort, cb)
-        }) : read(abort, cb)
+        _read ? _read(abort, getAbortCb(read, abort, cb)) : read(abort, cb)
       }
       else if(_read) nextChunk()
       else nextStream()
@@ -20,10 +29,7 @@ module.exports = function flatten () {
         _read(null, function (err, data) {
           if (err) {
             if (err === true) nextStream()
-            else read(true, function(abortErr) {
-              // TODO: what do we do with the abortErr?
-              cb(err)
-            })
+            else read(true, getChunkErrorCb(err, cb))
           }
           else cb(null, data)
         })