In my performance testing w/ [JITDB](https://github.com/ssb-ngi-pointer/jitdb), this change significantly reduced the time `flatten` spends at the top of the stack.

Please let me know if you need any changes to this PR before merging.