@Barbarrosa 2021-03-28
@noffle @dominictarr I added a benchmark to help measure the performance improvement. This change performs better in the benchmark than the original.

Also FYI the `Array.isArray` check was removed b/c it's redundant w/ the `'object' == typeof stream` check.