@Barbarrosa 2021-03-29
I encapsulated the abort/error case anonymous functions in static functions so they're only created when an error/abort is actually needed. This should help optimize the vast majority of cases, and it performs better in the new benchmark.