@Barbarrosa 2021-04-01
@mixmix I noticed `pull-stream` because it featured prominently when I generated flame graphs in JITDB using `0x -- node --expose-gc ./benchmark/index.js noCreate`. `flatten` is indicated by the cyan sections below.
![pull_stream_flatten_usages](https://user-images.githubusercontent.com/4523460/113239070-cbe07c80-925e-11eb-81ec-8dbeb831de4c.png)

The JITDB differences aren't that great, but I suspect it could improve performance more than expected given this library is used elsewhere in Manyverse.

<details>

<summary>JITDB Benchmark - Before</summary>

```
TAP version 13
# core indexes
ok 1 duration: 10ms
# query one huge index (first run)
ok 2 duration: 828ms
# query one huge index (second run)
ok 3 duration: 232ms
# count one huge index (third run)
ok 4 duration: 1ms
# create an index twice concurrently
ok 5 duration: 267ms
# query three indexes (first run)
ok 6 duration: 393ms
# query three indexes (second run)
ok 7 duration: 204ms
# load two indexes concurrently
ok 8 duration: 358ms
# paginate big index with small pageSize
ok 9 duration: 3954ms
# paginate big index with big pageSize
ok 10 duration: 186ms
# query a prefix map (first run)
ok 11 duration: 164ms
# query a prefix map (second run)
ok 12 duration: 10ms

1..12
# tests 12
# pass  12

# ok
```

</details>

<details>

  <summary>JITDB Benchmark - After</summary>

```
TAP version 13
# core indexes
ok 1 duration: 9ms
# query one huge index (first run)
ok 2 duration: 739ms
# query one huge index (second run)
ok 3 duration: 229ms
# count one huge index (third run)
ok 4 duration: 1ms
# create an index twice concurrently
ok 5 duration: 279ms
# query three indexes (first run)
ok 6 duration: 402ms
# query three indexes (second run)
ok 7 duration: 188ms
# load two indexes concurrently
ok 8 duration: 349ms
# paginate big index with small pageSize
ok 9 duration: 3943ms
# paginate big index with big pageSize
ok 10 duration: 190ms
# query a prefix map (first run)
ok 11 duration: 186ms
# query a prefix map (second run)
ok 12 duration: 10ms

1..12
# tests 12
# pass  12

# ok
```

</details>

The benchmark I wrote specifically to test how fast `flatten` works in this library does show a remarkable improvement.

Before:
```
$ node ./benchmarks/flatten.js 
flattenBench*100: 1.198s
79253.04
```

After:
```
$ node ./benchmarks/flatten.js 
flattenBench*100: 1.001s
-4014.88
```

The [toPullStream operator](https://github.com/ssb-ngi-pointer/jitdb#all-operators) and [low-level live function](https://github.com/ssb-ngi-pointer/jitdb#liveoperation-cb) in JITDB depend on `flatten`, and I suspect by their name and usage that they could be processing large amounts of data. [A brief search](https://gitlab.com/search?utf8=%E2%9C%93&search=toPullStream&group_id=&project_id=6814019&scope=&search_code=true&snippets=false&repository_ref=master&nav_source=navbar) confirms that Manyverse uses these functions.