@staltz 2021-04-01
@Barbarrosa Given [this comment](https://github.com/pull-stream/pull-stream/pull/130#discussion_r605571399), I don't see flatten playing a big role in performance, and the benchmarks show that the numbers could be within the normal fluctuation range, so not a significant change.

I think this PR is still useful, but I'll make some few suggestions in order to keep the readability (which is important unless the speed up is significant like >=20%). This PR contains some nice clean up and the addition of tests and benchmarks.