@Barbarrosa 2021-04-02
Here are the benchmark results after reverting the function extraction. It looks like `flatten` still keeps most of the performance gain.

```
$ node ./benchmarks/flatten.js 
flattenBench*100: 1.025s
29183.12
```