@Barbarrosa 2021-04-03
@staltz You're welcome! I opened up [a separate PR](https://github.com/pull-stream/pull-stream/pull/131) with some benchmark improvements.

Running that version against this PR yielded the following results.

Master:
```
flatten x 3,562,641 ops/sec ±2.13% (42 runs sampled)
Heap Change - max: 18.8 MB, min:-13.8 MB, mean:71 kB, std dev:7.85 MB
```

This PR:
```
flatten x 3,701,812 ops/sec ±1.30% (45 runs sampled)
Heap Change - max: 19 MB, min:-14.7 MB, mean:106 kB, std dev:8.67 MB
```